+++ mango/apps/conference/migrations/0001_initial.py
@@ -1,0 +1,44 @@
+# -*- coding: utf-8 -*-
+import datetime
+from south.db import db
+from south.v2 import SchemaMigration
+from django.db import models
+
+
+class Migration(SchemaMigration):
+
+    def forwards(self, orm):
+        # Adding model 'Conference'
+        db.create_table('conference_conference', (
+            ('id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
+            ('name', self.gf('django.db.models.fields.CharField')(max_length=100)),
+            ('starts', self.gf('django.db.models.fields.DateTimeField')()),
+            ('ends', self.gf('django.db.models.fields.DateTimeField')()),
+            ('description', self.gf('django.db.models.fields.TextField')()),
+            ('is_enabled', self.gf('django.db.models.fields.BooleanField')(default=False)),
+            ('created', self.gf('django.db.models.fields.DateTimeField')(auto_now_add=True, blank=True)),
+            ('updated', self.gf('django.db.models.fields.DateTimeField')(auto_now=True, null=True, blank=True)),
+        ))
+        db.send_create_signal('conference', ['Conference'])
+
+
+    def backwards(self, orm):
+        # Deleting model 'Conference'
+        db.delete_table('conference_conference')
+
+
+    models = {
+        'conference.conference': {
+            'Meta': {'object_name': 'Conference'},
+            'created': ('django.db.models.fields.DateTimeField', [], {'auto_now_add': 'True', 'blank': 'True'}),
+            'description': ('django.db.models.fields.TextField', [], {}),
+            'ends': ('django.db.models.fields.DateTimeField', [], {}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'is_enabled': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
+            'starts': ('django.db.models.fields.DateTimeField', [], {}),
+            'updated': ('django.db.models.fields.DateTimeField', [], {'auto_now': 'True', 'null': 'True', 'blank': 'True'})
+        }
+    }
+
+    complete_apps = ['conference']